requirements.txt
streamlink>=6.9.0
ffmpeg-python>=0.2.0
groq>=0.9.0
python-telegram-bot>=21.6
pydantic>=2.8.2
pydantic-settings>=2.4.0
aiofiles>=23.2.1
python-dotenv>=1.0.1

# ffmpeg must be installed separately via the OS package manager.